backend/app/pyproject.toml
[tool.poetry]
name = "psl-gateway"
version = "1.0.1b1"
description = ""
authors = ["a-ok123 <alexey@pastel.network>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.109"}
uvicorn = {extras = ["standard"], version = "^0.27.0"}
celery = "^5.2.7"
flower = "^2.0.0"
httpx = "^0.26.0"
aiofiles = "^23.1.0"
SQLAlchemy = "^2.0"
psycopg2-binary = "^2.9.5"
alembic = "^1.12"
tenacity = "^8.0.1"
pydantic = "^2.4"
passlib = {extras = ["bcrypt"], version = "^1.7"}
emails = "^0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
kombu = "^5.2.4"
redis = "^5.0.0"
ipfshttpclient = "^0.7.0"
zstd = "^1.5.4"
python-redis-cache = "^3.0"
redlock = "^1.2.0"
boto3 = "^1.34"
gunicorn = "^21.2.0"
pillow = "^10.0"
aiohttp = "^3.7.4"
secp256k1 = "^0.14.0"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
